+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 * (feature) Add `ClientOnlyPortal`.
 * (feature) Add Cookiebot hook.
+* (feature) Add `<Cookiebot>` snippet component.
 
 
 2.0.2